add tests for error handling and rename fixture
